Conflicts:
	cocos/2d/CCActionManager.h
	cocos/2d/CCClippingRectangleNode.h
	cocos/2d/CCGrid.h
	cocos/2d/CCLabel.h
	cocos/2d/CCMenuItem.h
	cocos/2d/CCNode.h
	cocos/2d/CCNodeGrid.h
	cocos/2d/CCScene.h
	cocos/3d/CCAnimate3D.h
	cocos/3d/CCSkybox.h
	cocos/3d/CCSprite3D.h
	cocos/3d/CCTerrain.h
	cocos/base/CCConfiguration.h
	cocos/base/CCDirector.h
	cocos/base/CCEventDispatcher.h
	cocos/base/CCUserDefault.h
	cocos/base/ccConfig.h
	cocos/base/ccUtils.h
	cocos/editor-support/cocostudio/ActionTimeline/CCFrame.h
	cocos/network/SocketIO.h
	cocos/physics3d/CCPhysics3DConstraint.h
	cocos/physics3d/CCPhysics3DObject.h
	cocos/physics3d/CCPhysics3DShape.h
	cocos/physics3d/CCPhysics3DWorld.h
	cocos/physics3d/CCPhysicsSprite3D.h
	cocos/platform/CCApplicationProtocol.h
	cocos/platform/CCDevice.h
	cocos/platform/CCFileUtils.h
	cocos/platform/android/CCEnhanceAPI-android.h
	cocos/ui/UICheckBox.h
	cocos/ui/UIEditBox/UIEditBox.h
	cocos/ui/UIPageView.h
	cocos/ui/UIWidget.h
	extensions/assets-manager/AssetsManagerEx.h
	extensions/assets-manager/Downloader.h
	tests/cpp-tests/Resources/ccs-res
	tools/bindings-generator
	tools/cocos2d-console

+++ cocos/platform/ios/CCApplication-ios.mm
@@ -57,7 +57,7 @@ int Application::run()
     return 0;
 }
 
-void Application::setAnimationInterval(double interval)
+void Application::setAnimationInterval(float interval)
 {
     [[CCDirectorCaller sharedDirectorCaller] setAnimationInterval: interval ];
 }
