+++ extensions/Particle3D/PU/CCPUOnQuotaObserver.cpp
@@ -63,7 +63,7 @@ void PUOnQuotaObserver::postUpdateObserver(float deltaTime)
         }
 
         //_result = particleTechnique->getNumberOfEmittedParticles(_particleTypeToObserve) >= quota;
-        _result = static_cast<PUParticleSystem3D *>(_particleSystem)->getAliveParticleCount() >= quota;
+        _result = static_cast<PUParticleSystem3D *>(_particleSystem)->getAliveParticleCount() >= (int)quota;
     }
     else
     {
@@ -77,7 +77,7 @@ void PUOnQuotaObserver::postUpdateObserver(float deltaTime)
             + static_cast<PUParticleSystem3D *>(_particleSystem)->getEmittedEmitterQuota()
             + static_cast<PUParticleSystem3D *>(_particleSystem)->getEmittedSystemQuota();
         //_result = particleTechnique->getNumberOfEmittedParticles() >= quota;
-        _result = static_cast<PUParticleSystem3D *>(_particleSystem)->getAliveParticleCount() >= quota;
+        _result = static_cast<PUParticleSystem3D *>(_particleSystem)->getAliveParticleCount() >= (int)quota;
     }
 }
